segregate different transformers
