+++ data/__init__.py
@@ -1,7 +1,7 @@
 import json
 from pathlib import Path
 
-from data.transformer import CITY_MAP_FILE
+from services.city_extractor_service import CITY_MAP_FILE
 
 
 def get_services(city: str) -> dict:
