+++ jobs.py
@@ -1,11 +1,13 @@
 from core import rq
-from data.transformer import ServiceTransformer
-from services.city_extractor_service import generate_code_name_mapper, fetch_city, NETWORK_WITH_CITY_FILE, CITY_MAP_FILE
+from data.transformers.name_transformer import NameTransformer
+from data.transformers.service_transformer import ServiceTransformer
+from services.city_extractor_service import fetch_city, NETWORK_WITH_CITY_FILE, \
+    CITY_MAP_FILE, TELECOM_FILE, CODE_NAME_FILE
 
 
 @rq.job(timeout=3600)
 def get_city_for_networks():
-    generate_code_name_mapper()
+    NameTransformer(input_file=TELECOM_FILE, output_file=CODE_NAME_FILE).transform()
     fetch_city()
     transformer = ServiceTransformer(input_file=NETWORK_WITH_CITY_FILE, output_file=CITY_MAP_FILE)
     transformer.transform()
